+++ Sources/Player/Common/Data/ProductType.swift
@@ -7,12 +7,11 @@ public typealias PlayerProductType = ProductType
 public enum ProductType: Codable, Equatable {
 	case TRACK
 	case VIDEO
-	case BROADCAST
 	case UC(url: URL)
 
 	func quality(given playbackContext: PlaybackContext) -> String {
 		switch self {
-		case .TRACK, .BROADCAST, .UC:
+		case .TRACK, .UC:
 			playbackContext.audioQuality?.rawValue ?? "AQ N/A"
 		case .VIDEO:
 			playbackContext.videoQuality?.rawValue ?? "VQ N/A"
@@ -21,7 +20,7 @@ public enum ProductType: Codable, Equatable {
 
 	func quality(given metadata: Metadata) -> String {
 		switch self {
-		case .TRACK, .BROADCAST, .UC:
+		case .TRACK, .UC:
 			metadata.audioQuality?.rawValue ?? "AQ N/A"
 		case .VIDEO:
 			metadata.videoQuality?.rawValue ?? "VQ N/A"
@@ -32,7 +31,6 @@ public enum ProductType: Codable, Equatable {
 		switch self {
 		case .TRACK: "TRACK"
 		case .VIDEO: "VIDEO"
-		case .BROADCAST: "BROADCAST"
 		case .UC: "UC"
 		}
 	}
@@ -43,7 +41,7 @@ public enum ProductType: Codable, Equatable {
 extension ProductType: CustomStringConvertible {
 	public var description: String {
 		switch self {
-		case .TRACK, .VIDEO, .BROADCAST: rawValue
+		case .TRACK, .VIDEO: rawValue
 		case let .UC(url): "\(rawValue) (\(url))"
 		}
 	}