+++ Sources/Player/Common/Event/NotificationsHandler.swift
@@ -43,24 +43,6 @@ final class NotificationsHandler {
 		}
 	}
 
-	func djSessionStarted(with metadata: DJSessionMetadata) {
-		queue.async {
-			self.listener.djSessionStarted(metadata)
-		}
-	}
-
-	func djSessionEnded(with reason: DJSessionEndReason) {
-		queue.async {
-			self.listener.djSessionEnded(with: reason)
-		}
-	}
-
-	func djSessionTransitioned(to transition: DJSessionTransition) {
-		queue.async {
-			self.listener.djSessionTransitioned(to: transition)
-		}
-	}
-
 	func offliningStarted(for mediaProduct: MediaProduct) {
 		queue.async {
 			self.offlineEngineListener?.offliningStarted(for: mediaProduct)