+++ Sources/Player/PlaybackEngine/Internal/AVQueuePlayer/AVPlayerAsset.swift
@@ -23,26 +23,3 @@ class AVPlayerAsset: Asset {
 	}
 }
 
-// MARK: - LiveAVPlayerAsset
-
-final class LiveAVPlayerAsset: AVPlayerAsset {
-	required init(
-		with player: AVQueuePlayerWrapper,
-		loudnessNormalizationConfiguration: LoudnessNormalizationConfiguration,
-		_ contentKeySession: AVContentKeySession?,
-		and contentKeySessionDelegate: AVContentKeySessionDelegate?
-	) {
-		super.init(
-			with: player,
-			loudnessNormalizationConfiguration: loudnessNormalizationConfiguration,
-			contentKeySession,
-			and: contentKeySessionDelegate
-		)
-	}
-
-	override func setAssetPosition(_ playerItem: AVPlayerItem) {
-		if let date = playerItem.currentDate() {
-			assetPosition = date.timeIntervalSince1970
-		}
-	}
-}