+++ Sources/Player/PlaybackEngine/Internal/Broadcasting/DJProducerListener.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol DJProducerListener: AnyObject {
-	func djSessionStarted(metadata: DJSessionMetadata)
-	func djSessionEnded(reason: DJSessionEndReason)
-}