+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Removed broadcast/DJ functionality from the Player module (Player)
+
 ## [0.7.7] - 2025-10-14
 
 ### Changed